Fix warnings in rc_pxc_util

+++ windows/RealityCapLib/rc_pxc_util.h
@@ -40,9 +40,9 @@ static rc_Vector rc_convertAcceleration(const imu_sample_t *s)
     rc_Vector acceleration_m__s2;
     //windows gives acceleration in g-units, so multiply by standard gravity in m/s^2
     //Also transform axes to be inline with RC definition
-    acceleration_m__s2.x = -s->data[1] * 9.80665;
-    acceleration_m__s2.y = s->data[0] * 9.80665;
-    acceleration_m__s2.z = -s->data[2] * 9.80665;
+    acceleration_m__s2.x = -s->data[1] * 9.80665f;
+    acceleration_m__s2.y = s->data[0] * 9.80665f;
+    acceleration_m__s2.z = -s->data[2] * 9.80665f;
     return acceleration_m__s2;
 }
 
@@ -51,8 +51,8 @@ static rc_Vector rc_convertAngularVelocity(const imu_sample_t *s)
     rc_Vector angular_velocity_rad__s;
     //windows gives angular velocity in degrees per second
     //Axes are different than accelerometer axes, and happen to be same as RC
-    angular_velocity_rad__s.x = s->data[0] * M_PI / 180.;
-    angular_velocity_rad__s.y = s->data[1] * M_PI / 180.;
-    angular_velocity_rad__s.z = s->data[2] * M_PI / 180.;
+    angular_velocity_rad__s.x = (float)(s->data[0] * M_PI / 180.f);
+    angular_velocity_rad__s.y = (float)(s->data[1] * M_PI / 180.f);
+    angular_velocity_rad__s.z = (float)(s->data[2] * M_PI / 180.f);
     return angular_velocity_rad__s;
 }
